# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CSMS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /complaints:
    get:
      operationId: listComplaints
      tags: [complaints]
      summary: List complaints with filters
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: status
          in: query
          schema:
            type: string
        - name: type_id
          in: query
          schema:
            type: integer
        - name: priority
          in: query
          schema:
            type: string
        - name: assigned_to_id
          in: query
          schema:
            type: integer
        - name: customer_id
          in: query
          schema:
            type: integer
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Complaints list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintsListResponse"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Create a complaint
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaintRequest"
      responses:
        "201":
          description: Complaint created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintItem"
  /complaints/{id}:
    get:
      operationId: getComplaint
      tags: [complaints]
      summary: Get complaint detail
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Complaint retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintDetail"
  /complaints/{id}/status:
    put:
      operationId: updateComplaintStatus
      tags: [complaints]
      summary: Update complaint status
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplaintStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintItem"
  /complaints/{id}/transfer:
    post:
      operationId: transferComplaint
      tags: [complaints]
      summary: Transfer complaint to another user
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferComplaintRequest"
      responses:
        "200":
          description: Complaint transferred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /complaints/{id}/escalate:
    post:
      operationId: escalateComplaint
      tags: [complaints]
      summary: Escalate complaint
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EscalateComplaintRequest"
      responses:
        "200":
          description: Complaint escalated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /complaints/{id}/logs:
    get:
      operationId: getComplaintLogs
      tags: [complaints]
      summary: Get complaint activity logs
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Complaint logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintLogsListResponse"
  /complaints/{id}/feedback:
    post:
      operationId: addComplaintFeedback
      tags: [complaints]
      summary: Add customer feedback to a complaint
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedbackRequest"
      responses:
        "201":
          description: Feedback added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackItem"
...
  /analytics/complaints:
    get:
      operationId: getComplaintAnalytics
      tags: [analytics]
      summary: Get complaint analytics
      security:
        - bearerAuth: []
      parameters:
        - name: date_from
          in: query
          schema:
            type: string
        - name: date_to
          in: query
          schema:
            type: string
        - name: branch_id
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Complaint analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComplaintAnalytics"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 25
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - total
        - page
        - limit
        - totalPages
...
    ComplaintItem:
      type: object
      properties:
        id:
          type: integer
        customer_id:
          type: integer
        customer_name:
          type: string
        customer_phone:
          type: string
        branch_name:
          type: string
        product_id:
          type: integer
        product_name:
          type: string
        invoice_id:
          type: integer
        invoice_number:
          type: string
        type_id:
          type: integer
        type_name:
          type: string
        channel:
          type: string
        priority:
          type: string
        description:
          type: string
        status:
          type: string
        assigned_to_id:
          type: integer
        assigned_to_name:
          type: string
        created_at:
          type: string
      required:
        - id
        - customer_id
        - type_id
        - channel
        - priority
        - description
        - status
    ComplaintDetail:
      allOf:
        - $ref: "#/components/schemas/ComplaintItem"
        - type: object
          properties:
            fields_values:
              type: object
            images:
              type: array
              items:
                type: string
            logs:
              type: array
              items:
                $ref: "#/components/schemas/ComplaintLogItem"
            feedback:
              $ref: "#/components/schemas/FeedbackItem"
    ComplaintsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ComplaintItem"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - data
        - meta
    CreateComplaintRequest:
      type: object
      properties:
        customer_id:
          type: integer
        product_id:
          type: integer
        invoice_id:
          type: integer
        type_id:
          type: integer
        fields_values:
          type: object
        channel:
          type: string
        priority:
          type: string
        description:
          type: string
        images:
          type: array
          items:
            type: string
      required:
        - customer_id
        - type_id
        - channel
        - priority
        - description
    UpdateComplaintStatusRequest:
      type: object
      properties:
        status:
          type: string
        note:
          type: string
      required:
        - status
    TransferComplaintRequest:
      type: object
      properties:
        target_user_id:
          type: integer
        note:
          type: string
      required:
        - target_user_id
    EscalateComplaintRequest:
      type: object
      properties:
        note:
          type: string
    ComplaintLogItem:
      type: object
      properties:
        id:
          type: integer
        complaint_id:
          type: integer
        action:
          type: string
        user_id:
          type: integer
        user_name:
          type: string
        note:
          type: string
        timestamp:
          type: string
      required:
        - id
        - complaint_id
        - action
        - user_id
        - timestamp
    ComplaintLogsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ComplaintLogItem"
      required:
        - data
    FeedbackItem:
      type: object
      properties:
        id:
          type: integer
        complaint_id:
          type: integer
        rating:
          type: integer
        comment:
          type: string
        created_at:
          type: string
      required:
        - id
        - complaint_id
        - rating
    CreateFeedbackRequest:
      type: object
      properties:
        rating:
          type: integer
        comment:
          type: string
      required:
        - rating
...
    ComplaintAnalytics:
      type: object
      properties:
        by_type:
          type: array
          items:
            type: object
        by_priority:
          type: array
          items:
            type: object
        avg_resolution_by_type:
          type: array
          items:
            type: object
        top_products:
          type: array
          items:
            type: object
      required:
        - by_type
...