# lib/api-zod/src/generated/types/updateComplaintStatusRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CSMS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateComplaintStatusRequest {
  status: string;
  note?: string;
}